pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["auto_verifier*", "report_engine*"]

[project]
name = "cfd-harness-unified"
version = "0.1.0"
description = "统一 AI-CFD 知识编译器：Foam-Agent 执行引擎 + Notion 控制面"
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
    "notion-client>=2.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 80
